# src/swagger.yml
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Invoice Validation",
    "description" : "validates an xml invoice and returns a report of results",
    "contact" : {
      "email" : "z5358645@ad.unsw.edu.au"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/team_fudge/invoice_validation/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "tags" : [ {
    "name" : "invoice_validation",
    "description" : "verifies xml invoices"
  } ],
  "paths" : {
    "/invoice/verify/all" : {
      "get" : {
        "tags" : [ "invoice_validation" ],
        "description" : "gets results report from invoice validation with all 4 components",
        "responses" : {
          "200" : {
            "description" : "results report is returned"
          }
        }
      },
      "post" : {
        "tags" : [ "invoice_validation" ],
        "description" : "user posts invoice xml to be validated against all 4 components",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Invoice"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "invoice is of type xml"
          },
          "400" : {
            "description" : "empty file"
          }
        }
      }
    },
    "/invoice/verify/wellformedness" : {
      "get" : {
        "tags" : [ "invoice_validation" ],
        "description" : "gets results report from invoice validation on wellformedness",
        "responses" : {
          "200" : {
            "description" : "results report is returned"
          }
        }
      },
      "post" : {
        "tags" : [ "invoice_validation" ],
        "description" : "user posts invoice xml to be validated against wellformedness",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Invoice"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "invoice is of type xml"
          },
          "400" : {
            "description" : "empty file"
          }
        }
      }
    },
    "/invoice/verify/syntax" : {
      "get" : {
        "tags" : [ "invoice_validation" ],
        "description" : "gets results report from invoice validation on syntax",
        "responses" : {
          "200" : {
            "description" : "results report is returned"
          }
        }
      },
      "post" : {
        "tags" : [ "invoice_validation" ],
        "description" : "user posts invoice xml to be validated against syntax",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Invoice"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "invoice is of type xml"
          },
          "400" : {
            "description" : "empty file"
          }
        }
      }
    },
    "/invoice/verify/peppol" : {
      "get" : {
        "tags" : [ "invoice_validation" ],
        "description" : "gets results report from invoice validation on peppol",
        "responses" : {
          "200" : {
            "description" : "results report is returned"
          }
        }
      },
      "post" : {
        "tags" : [ "invoice_validation" ],
        "description" : "user posts invoice xml to be validated against peppol",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Invoice"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "invoice is of type xml"
          },
          "400" : {
            "description" : "empty file"
          }
        }
      }
    },
    "/invoice/verify/schema" : {
      "get" : {
        "tags" : [ "invoice_validation" ],
        "description" : "gets results report from invoice validation on schema",
        "responses" : {
          "200" : {
            "description" : "results report is returned"
          }
        }
      },
      "post" : {
        "tags" : [ "invoice_validation" ],
        "description" : "user posts invoice xml to be validated against schema",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Invoice"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "invoice is of type xml"
          },
          "400" : {
            "description" : "empty file"
          }
        }
      }
    },
    "/active" : {
      "get" : {
        "description" : "returns activity status of all routes in API",
        "responses" : {
          "200" : {
            "description" : "healthcheck returned"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Invoice" : {
        "type" : "object",
        "additionalProperties" : {
          "type" : "string"
        },
        "description" : "xml invoice to be validated"
      }
    }
  }
}